backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
click==8.1.8
diskcache==5.6.3
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.5
llama_cpp_python==0.3.6
MarkupSafe==3.0.2
numpy==2.2.2
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
simple-websocket==1.1.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
wsproto==1.2.0
